test(refactor): split tests in mulitple files
each with setup and teardown needs to be run in band or it will collide with other tests
