+++ __tests__/routes/t-e-s-t-layout.ts
@@ -1,0 +1,77 @@
+jest.useFakeTimers();
+import express, { Application } from 'express';
+import 'reflect-metadata';
+import request from 'supertest';
+import { Connection } from 'typeorm';
+import routes from '../../src/lib/routes';
+import {
+  closeTestingConnections,
+  createTestingConnections,
+  reloadTestingDatabases,
+} from '../test-utils';
+
+// ███████╗███████╗████████╗██╗   ██╗██████╗
+// ██╔════╝██╔════╝╚══██╔══╝██║   ██║██╔══██╗
+// ███████╗█████╗     ██║   ██║   ██║██████╔╝
+// ╚════██║██╔══╝     ██║   ██║   ██║██╔═══╝
+// ███████║███████╗   ██║   ╚██████╔╝██║
+// ╚══════╝╚══════╝   ╚═╝    ╚═════╝ ╚═╝
+
+describe('misc functions that need a DB', () => {
+  let app: Application;
+  let connections: Connection[];
+
+  beforeAll(async (done) => {
+    if (process.env.NODE_ENV !== 'test') {
+      throw new Error('We are not in the test env this is harmful tables will be dropped');
+    }
+    connections = await createTestingConnections();
+    done();
+  });
+  // beforeEach(async (done) => {
+  //   try {
+  //     await reloadTestingDatabases(connections);
+  //     done();
+  //   } catch (err) {
+  //     console.warn(err.message);
+  //     console.warn(err.stack);
+  //   }
+  // });
+  afterAll(async (done) => {
+    try {
+      await reloadTestingDatabases(connections);
+      await closeTestingConnections(connections);
+      done();
+    } catch (err) {
+      console.warn(err.message);
+      console.warn(err.stack);
+      throw err;
+    }
+  });
+
+  app = express();
+  app.use(express.json());
+  app.use(express.urlencoded({ extended: true }));
+  app.use('/api/v1/', routes);
+
+// ███████╗███████╗████████╗██╗   ██╗██████╗
+// ██╔════╝██╔════╝╚══██╔══╝██║   ██║██╔══██╗
+// ███████╗█████╗     ██║   ██║   ██║██████╔╝
+// ╚════██║██╔══╝     ██║   ██║   ██║██╔═══╝
+// ███████║███████╗   ██║   ╚██████╔╝██║
+// ╚══════╝╚══════╝   ╚═╝    ╚═════╝ ╚═╝
+
+// ██████╗  ██████╗ ███╗   ██╗███████╗
+// ██╔══██╗██╔═══██╗████╗  ██║██╔════╝
+// ██║  ██║██║   ██║██╔██╗ ██║█████╗
+// ██║  ██║██║   ██║██║╚██╗██║██╔══╝
+// ██████╔╝╚██████╔╝██║ ╚████║███████╗
+// ╚═════╝  ╚═════╝ ╚═╝  ╚═══╝╚══════╝
+
+  test('route get user by id', async (done) => {
+    const res = await request(app).get('/api/v1/users/');
+    expect(res.status).toBe(200);
+    expect(Array.isArray(res.body.data)).toBe(true);
+    done();
+  });
+});
